demo2

+++ FleetTracker/Models/TruckerLog.cs
@@ -2,12 +2,12 @@ namespace FleetTracker.Models
 {
     public class TruckerLog
     {
-        public uint ID {get; set; }
+        public int ID {get; set; }
         public virtual Trucker Trucker {get; set; }
-        public uint TimeStamp {get; set; }
+        public long TimeStamp {get; set; }
         public float Latitude {get; set; }
         public float Longitude {get; set; }
-        public byte Speed {get; set; }
+        public float Speed {get; set; }
         public float Acceleration {get; set; }
     }
 }
